Fix pause map serialization for time offsets
